backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.3
email-validator>=2.0.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Async Job Queue
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0

# Authentication
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.0
python-jose[cryptography]>=3.3.0

# YAML Processing
pyyaml>=6.0.0

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# LLM Providers (optional - install as needed)
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
boto3>=1.28.0

# Validation
jsonschema>=4.20.0

# Observability
structlog>=23.2.0
prometheus-client>=0.19.0
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-exporter-otlp>=1.23.0
opentelemetry-instrumentation-fastapi>=0.44b0
opentelemetry-instrumentation-requests>=0.44b0

# Secrets (optional backends)
google-cloud-secret-manager>=2.18.0
hvac>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.0
